Remove id from update method

+++ LibraryManagement/Data/Interface/IBookRepository.cs
@@ -9,10 +9,10 @@ namespace LibraryManagement.Data.Interface
     public interface IBookRepository
     {
         Task<IEnumerable<Book>> GetAllBooksAsync();
-        Task<Book> GetBookByIdAsync(int bookId);
+        Task<Book> GetBookByIdAsync(long bookId);
 //        Task<BookExtended> GetBookWithDetailsAsync(Guid BookId);
         Task CreateBookAsync(Book book);
         Task UpdateBookAsync(Book newBook);
-        Task DeleteBookAsync(int bookId);
+        Task DeleteBookAsync(long bookId);
     }
 }
