+++ LibraryManagement/Data/Interface/IUserRepository.cs
@@ -9,10 +9,10 @@ namespace LibraryManagement.Data.Interface
     public interface IUserRepository
     {
         Task<IEnumerable<User>> GetAllUsersAsync();
-        Task<User> GetUserByIdAsync(int userId);
+        Task<User> GetUserByIdAsync(long userId);
         Task<User> CreateUserAsync(User user);
         Task UpdateUserAsync(User newUser);
-        Task DeleteUserAsync(int userId);
+        Task DeleteUserAsync(long userId);
         User AuthenticateUser(String userName, String password);
     }
 }
